I deleted comments that shouldn't be there, I deleted test logs, I deleted code that wasn't working, I modularized the files in the middleware directory since it was necessary due to the number of files in the directory and with this the paths where the middlewares are used are modified since their import changes when the location directories in which the modularization was carried out are changed.
