+++ backend/app/middleware/monitor/validateMonitorDisable.middleware.js
@@ -1,6 +1,6 @@
-const { User, Farm, Module, ROLES } = require('../../models');
-const ApiResponse = require('../utils/apiResponse');
-const { ROLES: Role } = require('../enums/roles.enum');
+const { User, Farm, Module, ROLES } = require('../../../models');
+const ApiResponse = require('../../utils/apiResponse');
+const { ROLES: Role } = require('../../enums/roles.enum');
 
 class ValidateMonitorDisableMiddleware {
     constructor() {}
