+++ backend/app/middleware/monitor/validateMonitorReactivate.middleware.js
@@ -1,6 +1,6 @@
-const { User, Farm, Module } = require('../../models');
-const ApiResponse = require('../utils/apiResponse');
-const { ROLES: Role } = require('../enums/roles.enum');
+const { User, Farm, Module } = require('../../../models');
+const ApiResponse = require('../../utils/apiResponse');
+const { ROLES: Role } = require('../../enums/roles.enum');
 
 class ValidateMonitorReactivateMiddleware {
     constructor() {}
