+++ backend/app/routes/admin/farm.admin.route.js
@@ -15,7 +15,6 @@ const farmAdminService = new FarmAdminService();
 const farmAdminController = new FarmAdminController(farmAdminService);
 const validateRoleMiddleware = new ValidateRoleMiddleware();
 
-// Create a Farm
 router.post(
     '/',
     validateTokenMiddleware.validate.bind(validateTokenMiddleware),
@@ -24,7 +23,6 @@ router.post(
     (req, res) => farmAdminController.create(req, res)
 );
 
-// List all Farms
 router.get(
     '/',
     validateTokenMiddleware.validate.bind(validateTokenMiddleware),
@@ -33,7 +31,6 @@ router.get(
     (req, res) => farmAdminController.index(req, res)
 );
 
-// Get a Farm
 router.get(
     '/:id',
     validateTokenMiddleware.validate.bind(validateTokenMiddleware),
@@ -41,7 +38,6 @@ router.get(
     (req, res) => farmAdminController.show(req, res)
 );
 
-// Update a Farm
 router.put(
     '/:id',
     validateTokenMiddleware.validate.bind(validateTokenMiddleware),
@@ -50,7 +46,6 @@ router.put(
     (req, res) => farmAdminController.update(req, res)
 );
 
-// Delete a Farm
 router.delete(
     '/:id',
     validateTokenMiddleware.validate.bind(validateTokenMiddleware),
