+++ backend/app/routes/admin/user.admin.route.js
@@ -10,8 +10,8 @@ const UserService = require("../../services/shared/user.service");
 const Mailer = require('../../utils/Mailer');
 const { ROLES: Role } = require("../../enums/roles.enum");
 const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
-const ValidateUserCreationMiddleware = require("../../middleware/validateUserCreation.middleware");
-const ValidateUserDeleteMiddleware = require("../../middleware/validateUserDelete.middleware");
+const ValidateUserCreationMiddleware = require("../../middleware/user/validateUserCreation.middleware");
+const ValidateUserDeleteMiddleware = require("../../middleware/user/validateUserDelete.middleware");
 
 
 const mailer = new Mailer(process.env.RESEND_API_KEY);
