+++ backend/app/routes/module/measurement.route.js
@@ -4,7 +4,7 @@ const ValidateTokenMiddleware = require("../../middleware/validateToken.middlewa
 const BlackListService = require("../../services/shared/blacklist.service");
 const { validate } = require('../../middleware/validate.middleware');
 const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
-const ValidateUserAccessMiddleware = require("../../middleware/validateUserAccess.middleware");
+const ValidateUserAccessMiddleware = require("../../middleware/user/validateUserAccess.middleware");
 const { ROLES : Role } = require("../../enums/roles.enum");
 const MeasurementController = require('../../controllers/module/measurement.controller');
 const MeasurementService = require('../../services/module/measurement.service');
