+++ backend/app/routes/shared/module.route.js
@@ -5,7 +5,7 @@ const ModuleController = require("../../controllers/shared/module.controller");
 const ValidateTokenMiddleware = require("../../middleware/validateToken.middleware");
 const BlackListService = require("../../services/shared/blacklist.service");
 const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
-const ValidateUserAccessMiddleware = require("../../middleware/validateUserAccess.middleware");
+const ValidateUserAccessMiddleware = require("../../middleware/user/validateUserAccess.middleware");
 const { ROLES: Role } = require("../../enums/roles.enum");
 const {validate} = require("../../middleware/validate.middleware");
 const {validateModulePaginate, validateIndexModules} = require("../../validators/shared/module.validator");
