+++ backend/app/routes/temp/notification.route.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const notificationController = require('../../controllers/temp/fcm.controller');
 
-// POST route to send a notification to a specific device
 router.post('/send', notificationController.sendDirectNotification);
 
 module.exports = router;
